# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StillGood App API
  description: |
    API для приложения StillGood.

    ## Роли
    - **USER** — поиск, бронирование, заказы
    - **PARTNER** — управление боксами, аналитика
    - **ADMIN** — веб-панель модерации

    ## Аутентификация
    - Access Token: 30 минут
    - Refresh Token: 14 дней
    - JWT Bearer token в заголовке Authorization
  version: 1.0.0
  contact:
    name: StillGood Development Team
    email: team@stillgood.tech
  x-team:
    - role: Tech Lead, Android Developer
      email: dmitry@stillgood.tech
    - role: Backend Developer
      email: timofey@stillgood.tech
    - role: Backend Developer
      email: artem@stillgood.tech
servers:
  - url: https://api.stillgood.tech/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Development
...
paths:
...
  /partner/offers:
    get:
      tags: [Partner]
      summary: Предложения партнёра
      operationId: getPartnerOffers
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
      responses:
        '200':
          description: Список предложений
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OfferDetailDto'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '403':
          description: Партнер не одобрен (PARTNER_NOT_APPROVED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Partner]
      summary: Создание предложения
      operationId: createOffer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOfferRequest'
      responses:
        '201':
          description: Предложение создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferDetailDto'
        '403':
          description: Партнер не одобрен (PARTNER_NOT_APPROVED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /partner/offers/{id}:
    patch:
      tags: [Partner]
      summary: Обновление предложения
      operationId: updateOffer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOfferRequest'
      responses:
        '200':
          description: Обновлённые данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferDetailDto'
        '403':
          description: Не ваше предложение или партнер не одобрен (PARTNER_NOT_APPROVED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [ Partner ]
      summary: Удаление предложения
      operationId: deleteOffer
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Предложение успешно удалено
        '403':
          description: Попытка удаления чужого предложения или партнер не одобрен (PARTNER_NOT_APPROVED)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Предложение не найдено
...
  /offers/{id}:
    get:
      tags: [Offers]
      summary: Детали предложения
      operationId: getOfferById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Данные предложения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OfferDetailDto'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ==================== CATEGORIES ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  schemas:
...
    RestaurantShortDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        phone:
          type: string
          nullable: true
...
    OfferCategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    OfferDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        discount:
          type: integer
          description: Процент скидки
        imageUrl:
          type: string
          nullable: true
        quantityAvailable:
          type: integer
        quantityTotal:
          type: integer
        pickupStart:
          type: string
          format: date-time
        pickupEnd:
          type: string
          format: date-time
        isActive:
          type: boolean
        category:
          $ref: '#/components/schemas/OfferCategoryDto'
        restaurant:
          $ref: '#/components/schemas/RestaurantShortDto'
    CreateOfferRequest:
      type: object
      required: [title, price, originalPrice, quantity, pickupStart, pickupEnd, categoryId]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        quantity:
          type: integer
        pickupStart:
          type: string
          format: date-time
        pickupEnd:
          type: string
          format: date-time
        categoryId:
          type: string
          format: uuid
        imageUrl:
          type: string
    UpdateOfferRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        quantity:
          type: integer
        isActive:
          type: boolean
        pickupStart:
          type: string
          format: date-time
        pickupEnd:
          type: string
          format: date-time
        categoryId:
          type: string
          format: uuid
        imageUrl:
          type: string

    # ==================== ORDERS ====================
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Человекочитаемое описание
...